LayoutTests/css3/scroll-snap/resnap-after-layout-expected.txt

layout trigger
PASS window.scrollY is 1200
PASS successfullyParsed is true

TEST COMPLETE
